Stage 2 with three steps: select exptype, select workflow, configure workflow; changed method getWorkflow(int id) to getWorkflow(long id) in WorkflowHandler.java and WorkflowHandlerImpl.java

+++ ptb/trunk/application/src/main/java/eu/planets_project/tb/api/model/mockups/WorkflowHandler.java
@@ -31,6 +31,6 @@ public interface WorkflowHandler {
 	
 	public ExperimentWorkflow getExperimentWorkflow(long lWorkflowEntityID);
 	
-	public Workflow getWorkflow(int workflowID);
+	public Workflow getWorkflow(long workflowID);
 
 }
